server/requirements.txt
Flask
flask_migrate
flask_script
Flask-SQLAlchemy
psycopg2
